refactor: broke index.js into multiple files
